Renamed mock service
