+++ src/main/java/com/gearvalley/infrastructure/DefaultWorldPort.java
@@ -1,5 +1,6 @@
 package com.gearvalley.infrastructure;
 
+import com.gearvalley.domain.WorldPort;
 import org.springframework.stereotype.Service;
 
 @Service
